+++ common/src/main/java/dev/cudzer/cobblemonalphas/ModEvents.java
@@ -9,10 +9,7 @@ import com.cobblemon.mod.common.api.battles.model.actor.BattleActor;
 import com.cobblemon.mod.common.api.events.CobblemonEvents;
 import com.cobblemon.mod.common.api.events.battles.BattleStartedEvent;
 import com.cobblemon.mod.common.api.pokemon.stats.Stats;
-import com.cobblemon.mod.common.battles.dispatch.DispatchResult;
-import com.cobblemon.mod.common.battles.dispatch.DispatchResultKt;
 import com.cobblemon.mod.common.battles.pokemon.BattlePokemon;
-import com.cobblemon.mod.common.net.messages.client.battle.BattleMessagePacket;
 import com.cobblemon.mod.common.pokemon.Pokemon;
 import com.cobblemon.mod.common.util.LocalizationUtilsKt;
 
@@ -20,7 +17,6 @@ import kotlin.Unit;
 import kotlin.jvm.functions.Function0;
 import net.minecraft.ChatFormatting;
 import net.minecraft.network.chat.MutableComponent;
-import net.minecraft.server.level.ServerPlayer;
 
 public class ModEvents {
 
@@ -65,6 +61,7 @@ public class ModEvents {
                 .withStyle(ChatFormatting.RED, ChatFormatting.BOLD);
 
         // Recursive retry until actors exist
+        @SuppressWarnings("unchecked")
         Function0<Unit>[] retry = new Function0[1];
 
         retry[0] = () -> {